+++ Server/Persistence/ReliableStorage.cs
@@ -7,74 +7,84 @@ namespace Server.Persistence
 {
     public class ReliableStorage : IDisposable
     {
+        public const string StorageFileName = "datastore.bin";
+        private const string TempFileName = "datastore.tmp";
         private readonly string _backupPath;
 
-        protected class BlockInfo
-        {
-            public BlockInfo(long offset, long transactionId)
-            {
-                Offset = offset;
-                TransactionId = transactionId;
-            }
 
-            public long Offset { get; }
+        private Thread _backupInitThread;
+        private BackupStorage _backupStorage;
 
-            /// <summary>
-            /// The id of the last transaction that modified the block
-            /// </summary>
-            public long TransactionId { get; }
-        }
-
-       
-        protected string DataPath { get; }
-
-        public int CorruptedBlocks { get; private set; }
-        public ReliableStorage(IPersistentObjectProcessor objectProcessor,  string workingDirectory = null, string backupPath = null, bool isBackup = false)
+        public ReliableStorage(IPersistentObjectProcessor objectProcessor, string workingDirectory = null,
+            string backupPath = null, bool isBackup = false)
         {
             _backupPath = backupPath;
 
             // the backup path is an absolute path. The data path is relative to the working directory
-            DataPath = isBackup ? _backupPath: workingDirectory != null ? Path.Combine(workingDirectory, Constants.DataPath): Constants.DataPath;
+            DataPath = isBackup ? _backupPath :
+                workingDirectory != null ? Path.Combine(workingDirectory, Constants.DataPath) : Constants.DataPath;
 
             ObjectProcessor = objectProcessor;
             IsBackup = isBackup;
 
-            if (!Directory.Exists(DataPath))
-            {
-                Directory.CreateDirectory(DataPath);
-            }
+            if (!Directory.Exists(DataPath)) Directory.CreateDirectory(DataPath);
 
-            string fullPath = Path.Combine(DataPath, StorageFileName);
+            var fullPath = Path.Combine(DataPath, StorageFileName);
 
-            _storageStream = new FileStream(fullPath, FileMode.OpenOrCreate);
+            StorageStream = new FileStream(fullPath, FileMode.OpenOrCreate);
+        }
 
-           
+
+        protected string DataPath { get; }
+
+        public int CorruptedBlocks { get; private set; }
+
+        private IPersistentObjectProcessor ObjectProcessor { get; }
+        public bool IsBackup { get; }
+
+        public int InactiveBlockCount { get; private set; }
+
+        public long StorageSize { get; private set; }
+
+        public int BlockCount => BlockInfoByPrimaryKey.Count;
+
+        public ISet<string> Keys => new HashSet<string>(BlockInfoByPrimaryKey.Keys);
+
+        protected Dictionary<string, BlockInfo> BlockInfoByPrimaryKey { get; } = new Dictionary<string, BlockInfo>();
+
+        protected Stream StorageStream { get; private set; }
+
+        public void Dispose()
+        {
+            StorageStream.Dispose();
+
+            _backupStorage?.Dispose();
         }
 
         public void LoadPersistentData(bool useObjectProcessor = true)
         {
+            bool IsPrimaryStorage()
+            {
+                return !IsBackup && _backupPath != null;
+            }
+
             // initialize the backup storage in parallel
-            if (!IsBackup && _backupPath != null)
+            if (IsPrimaryStorage())
             {
                 _backupStorage = new BackupStorage(_backupPath);
 
-                _backupInitThread = new Thread(() =>
-                {
-                    _backupStorage.LoadPersistentData();
-                    
-                });
+                _backupInitThread = new Thread(() => { _backupStorage.LoadPersistentData(); });
 
                 _backupInitThread.Start();
             }
 
-            bool success = false;
+            var success = false;
 
-            bool needsRecovery = false;
+            var needsRecovery = false;
 
-            _storageStream.Seek(0, SeekOrigin.Begin);
+            StorageStream.Seek(0, SeekOrigin.Begin);
 
             while (!success)
-            {
                 try
                 {
                     LoadAll(useObjectProcessor);
@@ -86,14 +96,10 @@ namespace Server.Persistence
                     needsRecovery = true;
                     MarkInvalidBlocksAsDirty(e);
                 }
-            }
 
 
             // wait for the backup to be initialized
-            if (_backupPath != null)
-            {
-                _backupInitThread.Join();
-            }
+            if (IsPrimaryStorage()) _backupInitThread.Join();
 
             // The storage was repaired but blocks have been lost. If a backup is available retrieve the 
             // missing block from the backup
@@ -108,22 +114,20 @@ namespace Server.Persistence
                     var block = _backupStorage.ReadBlock(blockKey);
                     StoreBlock(block.RawData, block.PrimaryKey, block.LastTransactionId);
                 }
-
             }
-
         }
 
         private void MarkInvalidBlocksAsDirty(InvalidBlockException e)
         {
-            long nextOffset = FindNextBeginMarker(e.Offset + PersistentBlock.MinSize);
+            var nextOffset = FindNextBeginMarker(e.Offset + PersistentBlock.MinSize);
 
-            long size = nextOffset - e.Offset;
+            var size = nextOffset - e.Offset;
 
             var dirtyBlock = PersistentBlock.MakeDirtyBlock(size);
 
-            _storageStream.Seek(e.Offset, SeekOrigin.Begin);
+            StorageStream.Seek(e.Offset, SeekOrigin.Begin);
 
-            var writer = new BinaryWriter(_storageStream);
+            var writer = new BinaryWriter(StorageStream);
 
             dirtyBlock.Write(writer);
 
@@ -134,51 +138,45 @@ namespace Server.Persistence
 
 
         /// <summary>
-        /// For recovery tests only
+        ///     For recovery tests only
         /// </summary>
         /// <param name="primaryKey"></param>
         public void MakeCorruptedBlock(string primaryKey)
         {
-            if(_blockInfoByPrimaryKey.TryGetValue(primaryKey, out var info))
+            if (BlockInfoByPrimaryKey.TryGetValue(primaryKey, out var info))
             {
                 var block = new PersistentBlock();
-                _storageStream.Seek(info.Offset, SeekOrigin.Begin);
-                var reader = new BinaryReader(_storageStream);
+                StorageStream.Seek(info.Offset, SeekOrigin.Begin);
+                var reader = new BinaryReader(StorageStream);
                 block.Read(reader);
 
                 block.Hash = block.Hash + 1; // the hash will not match so the block is corrupted
-                _storageStream.Seek(info.Offset, SeekOrigin.Begin);
+                StorageStream.Seek(info.Offset, SeekOrigin.Begin);
 
-                var writer = new BinaryWriter(_storageStream);
+                var writer = new BinaryWriter(StorageStream);
                 block.Write(writer);
             }
         }
 
         private long FindNextBeginMarker(long offset)
         {
-            BinaryReader reader = new BinaryReader(_storageStream);
-            for (long curOffset = offset; ; curOffset++)
-            {
+            var reader = new BinaryReader(StorageStream);
+            for (var curOffset = offset;; curOffset++)
                 try
                 {
-                    _storageStream.Seek(curOffset, SeekOrigin.Begin);
+                    StorageStream.Seek(curOffset, SeekOrigin.Begin);
                     var marker = reader.ReadInt32();
-                    if (marker == PersistentBlock.BeginMarkerValue)
-                    {
-
-                        return curOffset;
-                    }
+                    if (marker == PersistentBlock.BeginMarkerValue) return curOffset;
                 }
                 catch (EndOfStreamException)
                 {
                     return curOffset;
                 }
-            }
         }
 
         private void LoadAll(bool useObjectProcessor)
         {
-            var reader = new BinaryReader(_storageStream);
+            var reader = new BinaryReader(StorageStream);
 
             var block = new PersistentBlock();
             long offset = 0;
@@ -187,66 +185,46 @@ namespace Server.Persistence
             // read all the blocks
             while (block.Read(reader))
             {
-
                 // get information for deleted blocks too as there may be an uncomplete delete transaction to 
                 // reprocess
                 if (block.BlockStatus == BlockStatus.Active || block.BlockStatus == BlockStatus.Deleted)
                 {
-                    _blockInfoByPrimaryKey[block.PrimaryKey] =
+                    BlockInfoByPrimaryKey[block.PrimaryKey] =
                         new BlockInfo(offset, block.LastTransactionId);
-                    
+
                     // only active blocks contain objects 
                     if (useObjectProcessor && block.BlockStatus == BlockStatus.Active)
-                    {
                         ObjectProcessor.Process(block.RawData);
-                    }
-                    
                 }
 
-                if(block.BlockStatus != BlockStatus.Active)
-                {
-                    InactiveBlockCount++;
-                }
+                if (block.BlockStatus != BlockStatus.Active) InactiveBlockCount++;
 
-                offset = (int)_storageStream.Position;
+                offset = (int) StorageStream.Position;
             }
 
-            if (useObjectProcessor)
-            {
-                ObjectProcessor.EndProcess();
-            }
+            if (useObjectProcessor) ObjectProcessor.EndProcess();
 
             StorageSize = offset;
         }
 
-        public void Dispose()
-        {
-            _storageStream.Dispose();
-
-            _backupStorage?.Dispose();
-        }
-
         /// <summary>
-        /// Remove dirty blocks thus compacting the storage
+        ///     Remove dirty blocks thus compacting the storage
         /// </summary>
         public void CleanStorage()
         {
-            if (File.Exists(TempFileName))
-            {
-                File.Delete(TempFileName);
-            }
+            if (File.Exists(TempFileName)) File.Delete(TempFileName);
 
-            string fullPath = Path.Combine(DataPath, StorageFileName);
-            string tempPath = Path.Combine(DataPath, TempFileName);
+            var fullPath = Path.Combine(DataPath, StorageFileName);
+            var tempPath = Path.Combine(DataPath, TempFileName);
 
             using (var tempStream = new FileStream(tempPath, FileMode.Create))
             {
                 var writer = new BinaryWriter(tempStream);
-                _storageStream.Dispose();
+                StorageStream.Dispose();
 
-                _storageStream = new FileStream(fullPath, FileMode.OpenOrCreate);
+                StorageStream = new FileStream(fullPath, FileMode.OpenOrCreate);
 
-                var reader = new BinaryReader(_storageStream);
+                var reader = new BinaryReader(StorageStream);
 
                 var block = new PersistentBlock();
                 long offset = 0;
@@ -254,18 +232,16 @@ namespace Server.Persistence
 
                 // read all the blocks
                 while (block.Read(reader))
-                {
                     if (block.BlockStatus == BlockStatus.Active)
                     {
-                        offset = (int) _storageStream.Position;
+                        offset = (int) StorageStream.Position;
                         block.Write(writer);
                     }
-                }
 
                 StorageSize = offset;
 
                 writer.Flush();
-                _storageStream.Dispose();
+                StorageStream.Dispose();
             }
 
             File.Delete(fullPath);
@@ -273,7 +249,7 @@ namespace Server.Persistence
             File.Move(tempPath, fullPath);
 
 
-            _storageStream = new FileStream(fullPath, FileMode.OpenOrCreate);
+            StorageStream = new FileStream(fullPath, FileMode.OpenOrCreate);
 
             InactiveBlockCount = 0;
         }
@@ -281,34 +257,30 @@ namespace Server.Persistence
 
         public void DeleteBlock(string primaryKey, int transactionId)
         {
-            if (_blockInfoByPrimaryKey.TryGetValue(primaryKey, out var blockInfo))
+            if (BlockInfoByPrimaryKey.TryGetValue(primaryKey, out var blockInfo))
             {
-                _storageStream.Seek(blockInfo.Offset, SeekOrigin.Begin);
+                StorageStream.Seek(blockInfo.Offset, SeekOrigin.Begin);
 
-                var reader = new BinaryReader(_storageStream);
+                var reader = new BinaryReader(StorageStream);
                 var block = new PersistentBlock();
                 block.Read(reader);
                 if (block.BlockStatus != BlockStatus.Active)
-                {
                     // if the same id ok. We are re executing a transaction that was not marked as Processed. 
                     // It happens if the server crashes during the update of persistence blocks. The transaction is simply played
                     // again when the server is restarted
                     if (block.LastTransactionId != transactionId)
-                    {
-                        throw new NotSupportedException($"Trying to delete an inactive block for primary key {primaryKey}");
-                    }
-                    
-                }
+                        throw new NotSupportedException(
+                            $"Trying to delete an inactive block for primary key {primaryKey}");
 
                 block.BlockStatus = BlockStatus.Deleted;
                 block.LastTransactionId = transactionId;
 
-                _storageStream.Seek(blockInfo.Offset, SeekOrigin.Begin);
-                var writer = new BinaryWriter(_storageStream);
+                StorageStream.Seek(blockInfo.Offset, SeekOrigin.Begin);
+                var writer = new BinaryWriter(StorageStream);
                 block.Write(writer);
                 writer.Flush();
 
-                _blockInfoByPrimaryKey.Remove(primaryKey);
+                BlockInfoByPrimaryKey.Remove(primaryKey);
 
                 InactiveBlockCount++;
 
@@ -321,7 +293,7 @@ namespace Server.Persistence
         }
 
         /// <summary>
-        /// Add a new bloc or update an existing one (depending if the primary key is already stored)
+        ///     Add a new bloc or update an existing one (depending if the primary key is already stored)
         /// </summary>
         /// <param name="data"></param>
         /// <param name="primaryKey"></param>
@@ -337,12 +309,12 @@ namespace Server.Persistence
                 UsedDataSize = data.Length
             };
 
-            if (_blockInfoByPrimaryKey.TryGetValue(primaryKey, out var blockInfo))
+            if (BlockInfoByPrimaryKey.TryGetValue(primaryKey, out var blockInfo))
             {
                 // load the old version of the block
-                _storageStream.Seek(blockInfo.Offset, SeekOrigin.Begin);
+                StorageStream.Seek(blockInfo.Offset, SeekOrigin.Begin);
 
-                var reader = new BinaryReader(_storageStream);
+                var reader = new BinaryReader(StorageStream);
 
                 var oldBlock = new PersistentBlock();
                 oldBlock.Read(reader);
@@ -365,7 +337,7 @@ namespace Server.Persistence
                     block.ReservedDataSize = (int) (block.UsedDataSize * 1.5);
                     StorageSize = WriteBlock(block, StorageSize);
 
-                    _blockInfoByPrimaryKey[primaryKey] = new BlockInfo(StorageSize, block.LastTransactionId);
+                    BlockInfoByPrimaryKey[primaryKey] = new BlockInfo(StorageSize, block.LastTransactionId);
                 }
             }
             else // a new object not already in the persistent storage
@@ -376,51 +348,43 @@ namespace Server.Persistence
                 var offset = StorageSize;
                 StorageSize = WriteBlock(block, StorageSize);
 
-                _blockInfoByPrimaryKey[primaryKey] = new BlockInfo(offset, block.LastTransactionId);
+                BlockInfoByPrimaryKey[primaryKey] = new BlockInfo(offset, block.LastTransactionId);
             }
 
             _backupStorage?.StoreBlock(data, primaryKey, transactionId);
         }
 
-        public const string StorageFileName = "datastore.bin";
-        private const string TempFileName = "datastore.tmp";
-
-        private IPersistentObjectProcessor ObjectProcessor { get; }
-        public bool IsBackup { get; }
-        private Stream _storageStream;
-
-
-        private readonly Dictionary<string, BlockInfo> _blockInfoByPrimaryKey = new Dictionary<string, BlockInfo>();
-        private Thread _backupInitThread;
-        private BackupStorage _backupStorage;
-
-        public int InactiveBlockCount { get; private set; }
-
-        public long StorageSize { get; private set; }
-
-        public int BlockCount => _blockInfoByPrimaryKey.Count;
-
-        public ISet<string> Keys => new HashSet<string>(_blockInfoByPrimaryKey.Keys);
-
-        protected Dictionary<string, BlockInfo> BlockInfoByPrimaryKey => _blockInfoByPrimaryKey;
-
-        protected Stream StorageStream => _storageStream;
-
 
         /// <summary>
-        /// Write a block and return the offset after the block
+        ///     Write a block and return the offset after the block
         /// </summary>
         /// <param name="block"></param>
         /// <param name="offset"></param>
         /// <returns></returns>
-        long WriteBlock(PersistentBlock block, long offset)
+        private long WriteBlock(PersistentBlock block, long offset)
         {
-            _storageStream.Seek(offset, SeekOrigin.Begin);
-            var writer = new BinaryWriter(_storageStream);
+            StorageStream.Seek(offset, SeekOrigin.Begin);
+            var writer = new BinaryWriter(StorageStream);
             block.Write(writer);
             writer.Flush();
 
-            return _storageStream.Position;
+            return StorageStream.Position;
+        }
+
+        protected class BlockInfo
+        {
+            public BlockInfo(long offset, long transactionId)
+            {
+                Offset = offset;
+                TransactionId = transactionId;
+            }
+
+            public long Offset { get; }
+
+            /// <summary>
+            ///     The id of the last transaction that modified the block
+            /// </summary>
+            public long TransactionId { get; }
         }
     }
 }
